requirements.txt
# Core dependencies
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.0,<7.0.0

# Web interface
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
websockets>=12.0,<14.0

# Web fetching tools
playwright>=1.40.0,<2.0.0
requests>=2.31.0,<3.0.0
